tagsoup parsing for Atom-like feeds
XMLFeed is intended to be used as a fallback, when a feed does not parse
as valid RSS or Atom, but may be close enough to extract useful data
from. But, so far only RSS tags have been searched for in XMLFeed.
Extend that to look for Atom tags.

Closes https://github.com/bergmark/feed/issues/18

This is fairly complete, and I've tested everything I added using
real-world Atom feeds.

A few things omitted which could be added later got ToDo comments, and
two existing ToDo comments were implemented by this patch.

The tagsoup Atom parsing looks for XML items in the Atom namespace, so
this should not result in any behavior changes when parsing malformed RSS
or other non-Atom feeds.

Note the use of findChild in eg, getFeedTitle. Since Atom reuses tags
like title in a feed and in items, this was done to avoid traversing too
deep while looking for the feed title.