Changing HTMLContent to wrap an Elment: #35
Update HTMLContent to wrap an XML Element instead of text. Since the
HTML in a node will have been parsed as XML at this point in time, we
can treat it the same way we treat XML. This should allow users to parse
out HTML inside of a feed using their favorite XML->HTML mapping.